First Commit
First commit to github, right now I am implementing a Quadtree to
replace the nasty linear search of the VehicleManager. Hopefully it will
speed everything up so that more than 200 vehicles can be onscreen at a
time. This speedup will also come in handy when non-vehicle objects are
added for other steering behaviors.